undo move reader tests to reader_tests.rs
