+++ src/config/reader.rs
@@ -180,7 +180,7 @@ impl ConfigReader {
 }
 
 #[cfg(test)]
-mod test_proto_config {
+mod test_reader {
     use std::collections::HashMap;
     use std::path::{Path, PathBuf};
     use std::sync::Arc;
@@ -193,6 +193,7 @@ mod test_proto_config {
 
     use crate::cache::InMemoryCache;
     use crate::config::reader::ConfigReader;
+    use crate::config::{Config, Script, ScriptOptions, Type};
     use crate::http::Response;
     use crate::target_runtime::TargetRuntime;
     use crate::{EnvIO, HttpIO};
@@ -333,5 +334,115 @@ mod test_proto_config {
         } else {
             None
         }
+    }
+    fn start_mock_server() -> httpmock::MockServer {
+        httpmock::MockServer::start()
+    }
+
+    #[tokio::test]
+    async fn test_all() {
+        let runtime = init_runtime();
+
+        let mut cfg = Config::default();
+        cfg.schema.query = Some("Test".to_string());
+        cfg = cfg.types([("Test", Type::default())].to_vec());
+
+        let server = start_mock_server();
+        let header_serv = server.mock(|when, then| {
+            when.method(httpmock::Method::GET).path("/bar.graphql");
+            then.status(200).body(cfg.to_sdl());
+        });
+
+        let mut json = String::new();
+        tokio::fs::File::open("examples/jsonplaceholder.json")
+            .await
+            .unwrap()
+            .read_to_string(&mut json)
+            .await
+            .unwrap();
+
+        let foo_json_server = server.mock(|when, then| {
+            when.method(httpmock::Method::GET).path("/foo.json");
+            then.status(200).body(json);
+        });
+
+        let port = server.port();
+        let files: Vec<String> = [
+            "examples/jsonplaceholder.yml", // config from local file
+            format!("http://localhost:{port}/bar.graphql").as_str(), // with content-type header
+            format!("http://localhost:{port}/foo.json").as_str(), // with url extension
+        ]
+        .iter()
+        .map(|x| x.to_string())
+        .collect();
+        let cr = ConfigReader::init(runtime);
+        let c = cr.read_all(&files).await.unwrap();
+        pretty_assertions::assert_eq!(
+            ["Post", "Query", "Test", "User"]
+                .iter()
+                .map(|i| i.to_string())
+                .collect::<Vec<String>>(),
+            c.types
+                .keys()
+                .map(|i| i.to_string())
+                .collect::<Vec<String>>()
+        );
+        foo_json_server.assert(); // checks if the request was actually made
+        header_serv.assert();
+    }
+
+    #[tokio::test]
+    async fn test_local_files() {
+        let runtime = init_runtime();
+
+        let files: Vec<String> = [
+            "examples/jsonplaceholder.yml",
+            "examples/jsonplaceholder.graphql",
+            "examples/jsonplaceholder.json",
+        ]
+        .iter()
+        .map(|x| x.to_string())
+        .collect();
+        let cr = ConfigReader::init(runtime);
+        let c = cr.read_all(&files).await.unwrap();
+        pretty_assertions::assert_eq!(
+            ["Post", "Query", "User"]
+                .iter()
+                .map(|i| i.to_string())
+                .collect::<Vec<String>>(),
+            c.types
+                .keys()
+                .map(|i| i.to_string())
+                .collect::<Vec<String>>()
+        );
+    }
+
+    #[tokio::test]
+    async fn test_script_loader() {
+        let runtime = init_runtime();
+
+        let cargo_manifest = std::env::var("CARGO_MANIFEST_DIR").unwrap();
+        let reader = ConfigReader::init(runtime);
+
+        let config = reader
+            .read(&format!(
+                "{}/examples/jsonplaceholder_script.graphql",
+                cargo_manifest
+            ))
+            .await
+            .unwrap();
+
+        let path = format!("{}/examples/scripts/echo.js", cargo_manifest);
+        let file = ScriptOptions {
+            src: String::from_utf8(
+                tokio::fs::read(&path)
+                    .await
+                    .context(path.to_string())
+                    .unwrap(),
+            )
+            .unwrap(),
+            timeout: None,
+        };
+        pretty_assertions::assert_eq!(config.server.script, Some(Script::File(file)),);
     }
 }
